# 8306075: Micro-optimize Enum.hashCode