@cl4es 2023-04-17
Why isn't `Enum::hashCode` not simply doing `return ordinal;`?