@olivergillespie 2023-04-17
> Why isn't `Enum::hashCode` simply doing `return ordinal;`?

See https://bugs.openjdk.org/browse/JDK-8050217