@cl4es 2023-04-17
> > Why isn't `Enum::hashCode` simply doing `return ordinal;`?
> 
> See https://bugs.openjdk.org/browse/JDK-8050217

Thanks! If there are apps where `Enum::hashCode` is performance sensitive then run-to-run stability may be a stronger argument than @stuart-marks et al argues it is here, though there might be some unspoken arguments about how it might affect iteration order et.c... 

Adding a cached field to `Enum` seem innocuous though, whether it disappears into a padding gap or not.