@RogerRiggs 2023-04-17
Can the detailed explanation be moved to a comment. The full description from the PR is included in every email and it makes it inconvenient to find the continuing comments.