@shipilev 2023-04-18
> I do have one comment: Since identity hash codes are typically reasonably well-conditioned, it is perfectly reasonable to recondition an occasional 0 by replacing it with 1, in order to store it in a stable variable for later reuse.

As comment in the code change says, this is already handled on Hotspot side: https://github.com/openjdk/jdk/blob/cc60f2ff3f16bdb04917e09cb87f09bd544f1f8b/src/hotspot/share/runtime/synchronizer.cpp#L826-L827 -- if only to disambiguate the "unset" IHC in the object header. So the stars have already aligned well for us to use `0` as uninitialized marker!